Adapting to pip updates
Dear Derek, 

Due to recent changes from pip 10, one line of code is outdated. The line of code throws an error message when anyone wants to install quantopian-tools api. An adjustment was made to allow users to continue to clone your repo and install quantopian-tools based on the recent pip changes. Thank you so much for creating this repository. 

Sincerely,

Mark Maroki